+++ sender.c
@@ -50,7 +50,7 @@ void dataAvailable(pa_stream *IGN(stream), size_t IGN(bytes), void *IGN(userdata
   send(udpSocket, &packet, sizeof(packet) - sizeof(packet.data) + available, 0);
 
   position += available;
-  printf("Data transmitted. Position now at: %lu\n", position);
+  printf("Data transmitted. Position now at: %llu\n", (long long unsigned int)position);
 
   if(pa_stream_drop(stream)) {
     fprintf(stderr, "Failed to acknowledge stream data: %s\n", pa_strerror(pa_context_errno(ctx)));
